Addendum.

+++ chia/server/start_introducer.py
@@ -9,6 +9,7 @@ from chia.server.start_service import Service, async_run
 from chia.util.chia_logging import initialize_logging
 from chia.util.config import load_config, load_config_cli
 from chia.util.default_root import DEFAULT_ROOT_PATH
+from chia.util.ints import uint16
 
 # See: https://bugs.python.org/issue29288
 "".encode("idna")
@@ -39,7 +40,7 @@ def create_introducer_service(
         service_name=SERVICE_NAME,
         server_listen_ports=[service_config["port"]],
         network_id=network_id,
-        advertised_port=advertised_port,
+        advertised_port=uint16(advertised_port),
         connect_to_daemon=connect_to_daemon,
     )
